LPS-93661 Merge tests.
